requirements.txt
Django==2.2.8
djangorestframework==3.10.3
django-filter==2.2.0
psycopg2-binary==2.8.4
django-extensions==2.2.5
django-extra-fields==2.0.2
drf-writable-nested==0.5.3
flake8